+++ cheese-server/src/main/java/com/cheese/service/impl/DishServiceImpl.java
@@ -62,7 +62,7 @@ public class DishServiceImpl extends ServiceImpl<DishMapper,Dish> implements Dis
     {
         Dish dish = new Dish();
         BeanUtils.copyProperties(dishDTO, dish);
-        dish.setMerchantId(BaseContext.getCurrentId());
+        dish.setMerchantId((Long)BaseContext.getCurrentId());
         //向菜品表添加一条数据
         dishMapper.insert(dish);
 
@@ -88,7 +88,7 @@ public class DishServiceImpl extends ServiceImpl<DishMapper,Dish> implements Dis
                 new Page<>(dishPageQueryDTO.getPage(), dishPageQueryDTO.getPageSize()),
                 new QueryWrapper<DishVO>()
                         .like(StringUtils.isNotEmpty(dishPageQueryDTO.getName()), "d.name", dishPageQueryDTO.getName())
-                        .eq(BaseContext.getCurrentId() != null, "d.merchant_id", BaseContext.getCurrentId())
+                        .eq((Long)BaseContext.getCurrentId() != null, "d.merchant_id", (Long)BaseContext.getCurrentId())
                         .eq(dishPageQueryDTO.getCategoryId() != null, "d.category_id", dishPageQueryDTO.getCategoryId())
                         .eq(dishPageQueryDTO.getStatus() != null, "d.status", dishPageQueryDTO.getStatus())
                         .orderByAsc("d.create_time")
